Add procedure `nf95_abort`
`nf95_abort` is meant to replace `handle_err`. Essentially,
`nf95_abort` is `handle_err` without the test on the error
status. Using `handle_err` is not optimal when we go directly to the C
interface: there is a double test on the error status in the most
common case: no error and ncerr absent. From the point of view of
performance, it is also better to test the error status before the
call. I think it is also a little better for clarity.