Remove GNUmakefile
It is confusing to have both GNUmakefile and `CMakeLists.txt` and it
makes the documentation confusing to have this alternative. Also it is
annoying to have to maintain two build systems. Also, if NetCDF95 is
not built with CMake then it becomes more difficult to build dependent
programs (like `max_diff_nc` and ml2pl) with CMake.