Switch to default install of all mod files
With the NVidia Fortran compiler, a program using a library needs all
the mod files from the library, not just the regrouping mod
file. Instead of adding the compiler ID of the NVidia Fortran compiler
to the list of compilers which have this behaviour, we switch the
default behaviour to installing all the mod files. The GNU compiler
becomes the only compiler for which we only need to install the
regrouping mod file.