Use a relative submodule URL
Because we want to be able to download the project from irene at TGCC,
which allows access to the Gitlab of IN2P3 but not to Github. So the
project is mirrored in the Gitlab of IN2P3. The downside is that a
fork needs to modify the URL of the submodule, or fork the
submodule repository too. I considered using the URL of cmake in the
Gitlab of IN2P3 but there is the problem of the certificate of Gitlab,
not accepted by Ubuntu >= 22.04.