Remove function `inq_numgrps`
We are reverting part of commit 4b6812c2. Following an idea of Yann
Meurdesoif. Instead of using an ad hoc C function `inq_numgrps` to
call `nc_inq_grps` with a null pointer as third argument, we declare a
second interface to `nc_inq_grps` in procedure `nf95_inq_grps`. This
is simpler: we get rid of an intermediary function and we do not need
a C compiler for NetCDF95.