Remove module `nc_constants`
The only constant with kind `c_int` which we need is `nc_noerr`. It
seems uselessly complex to define other constants with kind `c_int`
and then copy them to constants with default kind. So we just define
directly the constants `Nf95_ENOGRP` and `Nf95_ENOTVAR` and move the
definition of `nc_noerr` to module `nf95_constants`.