Trim `full_name`
This is debatable. Whether we trim or not, we should do the same in
the call to `nc_inq_grp_full_ncid` and in the call to `nf95_abort`. I
have encountered a run where there was an extra space at the end of
`full_name` so the call to `nc_inq_grp_full_ncid` failed but the
erroneous space could not be seen in the error message of
`nf95_abort`, so the error was not made clear by the message.